Extract src-tauri/src/pty.rs — the launch layer
CLAUDE.md's "four launch engines, one telemetry path" as one module. All five
spawners came, not just the three that open a PTY: spawn_ghostty and
spawn_external_terminal write the same --settings file and are the same decision
from the frontend's side (termEngine), so splitting them from spawn_claude would
put one launch path in two modules. available_terminals and open_terminal_here
came with them; find_ghostty, interactive_shell, task_shell and apply_utf8_locale
are their cfg-gated helpers, held back from platform.rs in the first slice for
exactly this.

Also: write_pty / resize_pty / kill_session (they address a PTY by session id)
and Resources / session_resources (it resolves a session's pid out of the same
map). One test, the login-shell argv check.

lib.rs 2,595 -> 1,939. Imports trimmed on both sides to what each still uses:
lib.rs keeps only ChildKiller + MasterPty from portable_pty (the Session struct)
and home_dir + norm_path from platform; pty.rs took native_pty_system,
CommandBuilder, PtySize, and augmented_path/ps_one/resolve_claude/sys_command.

One thing the compiler had to tell me: spawn_task names tasks::TaskSpec and
tasks::Exec, so pty.rs takes `use crate::tasks;` and keeps the qualification
rather than importing the types — the moved lines then stay byte-identical,
which is what ground rule 2 is for.

Verification:
- pty.rs vs the pre-slice lib.rs, line-sorted: 3 differences, all of them the
  import lines I trimmed after the cut. Every moved line identical.
- the remainder of lib.rs: 31 lines, each intended — mod pty;, the four import
  trims, and the 11 pty:: qualifications in generate_handler!.

cargo test: 69 passed, same as baseline. cargo check --tests clean, no
warnings. No TypeScript touched.
